refactor: define `xhc` module
